+++ src/settings.cpp
@@ -2,12 +2,12 @@
 void wps::setparams( const wps_parameters params )
 {
     require_auth( get_self() );
-    const name ram_payer = get_self();
+
     check( params.voting_interval >= DAY, "[voting_interval] must be equal or above 24 hours (86400)");
     check( params.deposit_required.symbol == CORE_SYMBOL, "[deposit_required] invalid CORE_SYMBOL");
     check( params.max_monthly_budget.symbol == CORE_SYMBOL, "[max_monthly_budget] invalid CORE_SYMBOL");
 
-    _settings.set( params, ram_payer );
+    _settings.set( params, get_self() );
 }
 
 [[eosio::action]]
@@ -37,19 +37,15 @@ void wps::check_contract_active()
 
 void wps::add_funding( const asset quantity )
 {
-    const name ram_payer = get_self();
-
     auto state = _state.get();
     state.available_funding += quantity;
-    _state.set( state, ram_payer );
+    _state.set( state, get_self() );
 }
 
 void wps::sub_funding( const asset quantity )
 {
-    const name ram_payer = get_self();
-
     auto state = _state.get();
     state.available_funding -= quantity;
-    _state.set( state, ram_payer );
+    _state.set( state, get_self() );
 }