+++ src/vote.cpp
@@ -1,5 +1,5 @@
 [[eosio::action]]
-void wps::vote( const eosio::name voter, const eosio::name proposal_name, const eosio::name vote )
+void wps::vote( const name voter, const name proposal_name, const name vote )
 {
     require_auth( voter );
 
@@ -9,17 +9,17 @@ void wps::vote( const eosio::name voter, const eosio::name proposal_name, const 
     // cannot vote during completed voting period phase
     check_voting_period_completed();
 
+    // voter must be an active producer with over 100 EOS in total votes
+    check_voter_eligible( voter );
+
     // update `votes` table
     update_vote( voter, proposal_name, vote );
-
-    // must be producer with over 100 EOS in vpay
-    check_voter_eligible( voter );
 
     // update `proposals::eligible` field for all active proposals
     update_eligible_proposals();
 }
 
-void wps::check_voter_eligible( const eosio::name voter )
+void wps::check_voter_eligible( const name voter )
 {
     eosiosystem::producers_table _producers( "eosio"_n, "eosio"_n.value );
     eosiosystem::global_state_singleton _gstate( "eosio"_n, "eosio"_n.value );
@@ -27,15 +27,13 @@ void wps::check_voter_eligible( const eosio::name voter )
     auto prod = _producers.get( voter.value, "[voter] must be registered as a producer" );
     auto gstate = _gstate.get();
 
-    const time_point last_24h = time_point(current_time_point() - time_point_sec(DAY));
-    const int64_t producer_per_vote_pay = int64_t((gstate.pervote_bucket * prod.total_votes) / gstate.total_producer_vote_weight);
+    check( prod.total_votes > 0, "[voter] must have votes");
 
-    check( producer_per_vote_pay > 0, "[voter] must have votes");
-    check( prod.last_claim_time >= last_24h, "[voter] must have claimed producer rewards within the last 24 hours");
+    const int64_t producer_per_vote_pay = int64_t((gstate.pervote_bucket * prod.total_votes) / gstate.total_producer_vote_weight);
     check( producer_per_vote_pay >= 1000000, "[voter] must be have a vpay of 100 EOS or above");
 }
 
-void wps::check_proposal_can_vote( const eosio::name proposal_name )
+void wps::check_proposal_can_vote( const name proposal_name )
 {
     auto proposals_itr = _proposals.find( proposal_name.value );
     check( proposals_itr != _proposals.end(), "[proposal_name] does not exists");
@@ -43,18 +41,7 @@ void wps::check_proposal_can_vote( const eosio::name proposal_name )
     check( proposals_itr->end > current_time_point(), "[proposal_name] has ended");
 }
 
-int16_t wps::calculate_total_net_votes( const std::map<eosio::name, eosio::name> votes )
-{
-    int16_t total_net_votes = 0;
-    for (std::pair<eosio::name, eosio::name> item : votes) {
-        const eosio::name vote = item.second;
-        if (vote == "yes"_n) total_net_votes += 1;
-        else if (vote == "no"_n) total_net_votes -= 1;
-    }
-    return total_net_votes;
-}
-
-void wps::update_vote( const eosio::name voter, const eosio::name proposal_name, const eosio::name vote )
+void wps::update_vote( const name voter, const name proposal_name, const name vote )
 {
     // validate vote
     auto votes_itr = _votes.find( proposal_name.value );
@@ -64,14 +51,10 @@ void wps::update_vote( const eosio::name voter, const eosio::name proposal_name,
     check( vote == "yes"_n || vote == "no"_n || vote == "abstain"_n, "[vote] invalid (ex: yes/no/abstain)");
 
     // update vote
-    _votes.modify( votes_itr, same_payer, [&]( auto& vote_row ) {
-        check(vote_row.votes[voter] != vote, "[vote] has not been modified");
-        vote_row.votes[voter] = vote;
-
-        // re-caculate votes
-        _proposals.modify( proposals_itr, same_payer, [&]( auto& proposals_row ) {
-            proposals_row.total_net_votes = calculate_total_net_votes( vote_row.votes );
-        });
+    _votes.modify( votes_itr, same_payer, [&]( auto& row ) {
+        check(row.votes[voter] != vote, "[vote] has not been modified");
+        row.votes[voter] = vote;
+        update_total_net_votes( proposal_name, row.votes );
     });
 }
 
@@ -126,4 +109,27 @@ std::map<int16_t, std::set<eosio::name>> wps::sort_proposals_by_net_votes( const
         proposals[itr->total_net_votes].insert(itr->proposal_name);
     }
     return proposals;
+}
+
+
+int16_t wps::calculate_total_net_votes( const std::map<name, name> votes )
+{
+    int16_t total_net_votes = 0;
+    for (std::pair<name, name> item : votes) {
+        const name vote = item.second;
+        if (vote == "yes"_n) total_net_votes += 1;
+        else if (vote == "no"_n) total_net_votes -= 1;
+    }
+    return total_net_votes;
+}
+
+void wps::update_total_net_votes( const name proposal_name, const std::map<name, name> votes )
+{
+    auto proposal_itr = _proposals.find( proposal_name.value );
+    check( proposal_itr != _proposals.end(), "[update_total_net_votes::proposal_name] is not found");
+
+    // re-caculate votes
+    _proposals.modify( proposal_itr, same_payer, [&]( auto& row ) {
+        row.total_net_votes = calculate_total_net_votes( votes );
+    });
 }